Move balance queries to a separate module
